feat(r2): remove cloudinary backend + converge deployment config (#827 phase 7)
- Delete api/cloudinary/ (views, admin_views), api/cloudinary_cleanup.py,
  and their tests; drop the admin cloudinary-cleanup routes and exports.
- Inline _StreamingZipBuffer into auth/export_archive.py (its only
  remaining consumer).
- Deployment config: CLOUDINARY_* -> R2_* across .env examples, Helm
  configmap/deployments/values, render_values.sh, cd.yml vars, and the
  ci.yml chart-render checks.
- [SECURITY] CSP: drop upload-widget/api/widget cloudinary hosts, add
  the R2 CDN domain to img-src/media-src and the R2 storage endpoint to
  connect-src; res.cloudinary.com stays transitionally until
  migrate_assets_to_r2 runs in prod (removed in follow-up PR).
- Bazel: add crops.py to api_schema_lib (fixes ModuleNotFoundError in
  api_model_test/api_health_test) and add crops.py, r2.py, uploads/* to
  api_runtime_files so the OCI image ships the new modules.
- Type the presigned-endpoint resource config as a NamedTuple (mypy).
