+++ .env.production.example
@@ -21,13 +21,15 @@ APP_ORIGIN=
 # The frontend bundle also reads this variable to render the Google Login button.
 GOOGLE_OAUTH_CLIENT_ID=
 
-# Cloudinary image storage (optional — leave blank if not using)
-CLOUDINARY_CLOUD_NAME=
-CLOUDINARY_API_KEY=
-CLOUDINARY_API_SECRET=
-CLOUDINARY_UPLOAD_FOLDER=
-CLOUDINARY_PUBLIC_UPLOAD_FOLDER=
-CLOUDINARY_UPLOAD_PRESET=
+# Cloudflare R2 object storage (image/video/audio assets)
+# All five are required for uploads, the eager crop pipeline, and showcase
+# video hosting; leave blank to disable uploads (stored URLs still render).
+R2_ACCOUNT_ID=
+R2_ACCESS_KEY_ID=
+R2_SECRET_ACCESS_KEY=
+R2_BUCKET_NAME=
+# Public CDN base URL for the bucket (custom domain), e.g. https://media.potterdoc.com
+R2_PUBLIC_URL=
 
 # Invitation email delivery via Resend SMTP
 EMAIL_HOST=smtp.resend.com
