+++ api/BUILD.bazel
@@ -41,11 +41,8 @@ py_library(
         "piece/helpers.py",
         "piece/image_views.py",
         "piece/showcase_views.py",
-        "cloudinary/__init__.py",
-        "cloudinary/views.py",
-        "cloudinary/admin_views.py",
-        "cloudinary_cleanup.py",
         "crop_run_views.py",
+        "crops.py",
         "global_entries/__init__.py",
         "global_entries/logic.py",
         "global_entries/views.py",
@@ -100,7 +97,6 @@ py_library(
     name = "api_lib",
     srcs = [
         "admin.py",
-        "cloudinary_cleanup.py",
         "manual_tile_imports.py",
         "model_factories.py",
         "tasks.py",
@@ -199,11 +195,8 @@ filegroup(
         "piece/helpers.py",
         "piece/image_views.py",
         "piece/showcase_views.py",
-        "cloudinary/__init__.py",
-        "cloudinary/views.py",
-        "cloudinary/admin_views.py",
-        "cloudinary_cleanup.py",
         "crop_run_views.py",
+        "crops.py",
         "global_entries/__init__.py",
         "global_entries/logic.py",
         "global_entries/views.py",
@@ -220,6 +213,7 @@ filegroup(
         "model_factories.py",
         "models.py",
         "preferences.py",
+        "r2.py",
         "serializer_factories.py",
         "serializer_registry.py",
         "serializers.py",
@@ -231,6 +225,8 @@ filegroup(
         "task_views.py",
         "telemetry_views.py",
         "tasks.py",
+        "uploads/__init__.py",
+        "uploads/views.py",
         "urls.py",
         "utils.py",
         "views.py",
@@ -495,7 +491,6 @@ pytest_test(
     srcs = _CONFTEST + [
         "tests/test_admin_exports.py",
         "tests/test_admin_scope.py",
-        "tests/test_cloudinary_cleanup.py",
         "tests/test_r2_image_widget.py",
         "tests/test_management_commands.py",
         "tests/test_r2_uploads.py",
@@ -503,7 +498,6 @@ pytest_test(
     args = [
         "api/tests/test_admin_exports.py",
         "api/tests/test_admin_scope.py",
-        "api/tests/test_cloudinary_cleanup.py",
         "api/tests/test_r2_image_widget.py",
         "api/tests/test_management_commands.py",
         "api/tests/test_r2_uploads.py",
@@ -514,7 +508,6 @@ pytest_test(
     deps = [":api_schema_lib", ":api_admin_lib"] + _PYTEST_DEPS,
     expected_coverage = [
         "api/admin.py",
-        "api/cloudinary_cleanup.py",
         "api/uploads/views.py",
         "api/r2.py",
         "api/management/commands/*.py",
