+++ chart/glaze/values.yaml
@@ -28,11 +28,9 @@ appConfig:
   appOrigin: ""
   inviteLinkBaseUrl: ""
   googleOauthClientId: ""
-  cloudinaryCloudName: ""
-  cloudinaryUploadFolder: ""
-  cloudinaryUploadPreset: ""
-  cloudinaryPublicUploadFolder: ""
-  cloudinaryVideoUploadFolder: ""
+  r2AccountId: ""
+  r2BucketName: ""
+  r2PublicUrl: ""
   emailHost: ""
   emailPort: ""
   emailHostUser: ""
